Desktop/multimodal_steganography_project/app/encryption.py
# app/encryption.py
from Cryptodome.Cipher import AES
from Cryptodome.Util.Padding import pad, unpad
from Cryptodome.Random import get_random_bytes

key = b'your_secure_key_here_32bytes_long'

# Encryption function
def encrypt_message(message: str):
    cipher = AES.new(key, AES.MODE_CBC)
    ct_bytes = cipher.encrypt(pad(message.encode(), AES.block_size))
    return cipher.iv + ct_bytes  # Return the IV + ciphertext

# Decryption function
def decrypt_message(encrypted_message: bytes):
    iv = encrypted_message[:16]
    ct = encrypted_message[16:]
    cipher = AES.new(key, AES.MODE_CBC, iv)
    pt = unpad(cipher.decrypt(ct), AES.block_size)
    return pt.decode()
