Desktop/multimodal_steganography_project/requirements.txt
flask==2.0.1
pycryptodome==3.10.1
pillow==8.3.1
librosa==0.8.1
numpy==1.21.1
soundfile==0.10.3.post1